+++ system/autoware_state_monitor/config/autoware_state_monitor.planning_simulation.yaml
@@ -10,7 +10,7 @@
       "planning",
       "control",
       "vehicle",
-      "emergency_system"
+      "system"
     ]
 
 # Topic Config
@@ -47,18 +47,6 @@
           timeout: 0.0
           warn_rate: 0.0
           type: "sensor_msgs/msg/PointCloud2"
-
-        /perception/object_recognition/detection/objects:
-          module: "perception"
-          timeout: 1.0
-          warn_rate: 5.0
-          type: "autoware_perception_msgs/msg/DynamicObjectWithFeatureArray"
-
-        /perception/object_recognition/tracking/objects:
-          module: "perception"
-          timeout: 1.0
-          warn_rate: 5.0
-          type: "autoware_perception_msgs/msg/DynamicObjectArray"
 
         /perception/object_recognition/objects:
           module: "perception"
@@ -103,7 +91,7 @@
           type: "autoware_vehicle_msgs/msg/Steering"
 
         /system/emergency/control_cmd:
-          module: "emergency_system"
+          module: "system"
           timeout: 1.0
           warn_rate: 5.0
           type: "autoware_control_msgs/msg/ControlCommandStamped"
