+++ system/autoware_state_monitor/config/autoware_state_monitor.yaml
@@ -10,7 +10,7 @@
       "planning",
       "control",
       "vehicle",
-      "emergency_system"
+      "system"
     ]
 
     # Topic Config
@@ -70,18 +70,6 @@
           warn_rate: 5.0
           type: "geometry_msgs/msg/PoseStamped"
 
-        /perception/object_recognition/detection/objects:
-          module: "perception"
-          timeout: 1.0
-          warn_rate: 5.0
-          type: "autoware_perception_msgs/msg/DynamicObjectWithFeatureArray"
-
-        /perception/object_recognition/tracking/objects:
-          module: "perception"
-          timeout: 1.0
-          warn_rate: 5.0
-          type: "autoware_perception_msgs/msg/DynamicObjectArray"
-
         # Can be both with feature array or without
         # In prediction.launch.xml this is set to without
         /perception/object_recognition/objects:
@@ -131,7 +119,7 @@
           type: "autoware_vehicle_msgs/msg/Steering"
 
         /system/emergency/control_cmd:
-          module: "emergency_system"
+          module: "system"
           timeout: 1.0
           warn_rate: 5.0
           type: "autoware_control_msgs/msg/ControlCommandStamped"
